Add subcommand dispatcher in main, stub config subcommand
Splits cmd/server/main.go into a thin dispatcher plus per-subcommand
files (identity.go, config.go). The identity subcommand preserves all
existing behavior; legacy top-level flags (--reset-admin,
--rotate-jwt-key, --list-backups, --restore-backup, etc.) still route
to identity for backward compatibility, so existing systemd units and
scripts keep working unchanged.

The config subcommand is a stub that prints usage and errors with
"not yet implemented" — it exists so the dispatcher has a real
target and so subsequent milestones can wire in real behavior without
touching the dispatch logic again.

No functional change to identity. All 162 e2e checks still pass, plus
unit and integration tests.
